requirements.txt
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.2.2
tensorflow-cpu==2.16.1
tensorflow-datasets==4.9.2